fixed bug for Mobile Navbar User Actions.
