+++ src/containers/navbar/components/NavbarMobile/components/NavbarUserActions/assets/utils.js
@@ -1,1 +1,1 @@
-export auth from '../../../../../../../navigation/routes';
+export { auth } from '../../../../../../../navigation/routes';
